fix(be): remove duplicate Module entity and use CourseModule

+++ backend/src/main/java/org/eduspace/backend/dto/study_group/response/GroupMemberDTO.java
@@ -15,5 +15,5 @@ public class GroupMemberDTO {
     private String email;
     private String username;
     private String avatarUrl;
-    private Long totalExp;
+    private Integer totalExp;
 }
